Fix: Corregir tipo de retorno en InteresUtils.calcularProgreso
- Convertir explÃ­citamente Double a Float con toFloat()
- Resolver error 'Return type mismatch: expected Float, actual Double'
- Uso de 100.0 (Double) y luego conversiÃ³n a Float

+++ app/src/main/java/com/example/bsprestagil/utils/InteresUtils.kt
@@ -142,7 +142,7 @@ object InteresUtils {
     fun calcularProgreso(capitalOriginal: Double, capitalPendiente: Double): Float {
         if (capitalOriginal == 0.0) return 0f
         val capitalPagado = capitalOriginal - capitalPendiente
-        return ((capitalPagado / capitalOriginal) * 100f).coerceIn(0f, 100f)
+        return ((capitalPagado / capitalOriginal) * 100.0).toFloat().coerceIn(0f, 100f)
     }
 }
